+++ hc/accounts/tests/test_admin.py
@@ -13,7 +13,30 @@ class AccountsAdminTestCase(BaseTestCase):
 
     def test_it_shows_profiles(self) -> None:
         self.client.login(username="alice@example.org", password="password")
-
         r = self.client.get("/admin/accounts/profile/")
         self.assertContains(r, "alice@example.org")
         self.assertContains(r, "bob@example.org")
+
+    def test_it_escapes_emails_when_showing_profiles(self) -> None:
+        self.bob.email = "bob&friends@example.org"
+        self.bob.save()
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.get("/admin/accounts/profile/")
+        # The amperstand should be escaped
+        self.assertNotContains(r, "bob&friends@example.org")
+
+    def test_it_shows_projects(self) -> None:
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.get("/admin/accounts/project/")
+        self.assertContains(r, "Alices Project")
+        self.assertContains(r, "Default Project for bob@example.org")
+
+    def test_it_escapes_emails_when_showing_projects(self) -> None:
+        self.bob.email = "bob&friends@example.org"
+        self.bob.save()
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.get("/admin/accounts/project/")
+        # The amperstand should be escaped
+        self.assertNotContains(r, "bob&friends@example.org")